pyproject.toml
[project]
name = "hidden-in-plain-script"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.14,<3.15"
dependencies = [
    "plotly (>=6.6.0,<7.0.0)",
    "nnsight (>=0.6.2,<0.7.0)",
    "google-transliteration-api (>=1.0.3,<2.0.0)",
    "torch (>=2.10.0,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "anthropic (>=0.85.0,<0.86.0)",
    "openai (>=2.29.0,<3.0.0)",
    "datasets (>=4.8.2,<5.0.0)",
    "pydantic-ai-slim[anthropic,openai] (>=1.69.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[poetry.tools]
package-mode = false
